+++ packages/logger/lib/index.js
@@ -1,5 +1,5 @@
 "use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
 // @kaapi/logger
 // logger
+Object.defineProperty(exports, "__esModule", { value: true });
 //# sourceMappingURL=index.js.map