+++ packages/logger/src/index.ts
@@ -1,2 +1,12 @@
 // @kaapi/logger
 // logger
+
+export interface ILogger {
+    silly: (...args: unknown[]) => void;
+    debug: (...args: unknown[]) => void;
+    verbose: (...args: unknown[]) => void;
+    info: (...args: unknown[]) => void;
+    warn: (...args: unknown[]) => void;
+    error: (...args: unknown[]) => void;
+    fatal: (...args: unknown[]) => void;
+}